- Add ability to add rewards onto research branches and have users claim them manually - Adds Research Trees into Extension Tracker (run php artisan update-extension-tracker) - you may need to clear config cache for this!

+++ app/Services/ItemService.php
@@ -319,6 +319,7 @@ class ItemService extends Service
             if(DB::table('character_items')->where([['item_id', '=', $item->id], ['count', '>', 0]])->exists()) throw new \Exception("At least one character currently owns this item. Please remove the item(s) before deleting it.");
             if(DB::table('loots')->where('rewardable_type', 'Item')->where('rewardable_id', $item->id)->exists()) throw new \Exception("A loot table currently distributes this item as a potential reward. Please remove the item before deleting it.");
             if(DB::table('prompt_rewards')->where('rewardable_type', 'Item')->where('rewardable_id', $item->id)->exists()) throw new \Exception("A prompt currently distributes this item as a reward. Please remove the item before deleting it.");
+            if(DB::table('research_rewards')->where('rewardable_type', 'Item')->where('rewardable_id', $item->id)->exists()) throw new \Exception("A research branch currently distributes this item as a reward. Please remove the item before deleting it.");
             if(DB::table('shop_stock')->where('item_id', $item->id)->exists()) throw new \Exception("A shop currently stocks this item. Please remove the item before deleting it.");
 
             DB::table('items_log')->where('item_id', $item->id)->delete();
